+++ src/RestaurantDetail.jsx
@@ -1,6 +1,7 @@
 import MenuItems from './MenuItems';
+import ReviewContainer from './ReviewContainer';
 
-export default function RestaurantDetail({ restaurant }) {
+export default function RestaurantDetail({ reviews, restaurant }) {
   const { name, address, menuItems } = restaurant;
 
   return (
@@ -13,6 +14,7 @@ export default function RestaurantDetail({ restaurant }) {
       </p>
       <h3>메뉴</h3>
       <MenuItems menuItems={menuItems} />
+      <ReviewContainer reviews={reviews} />
     </div>
   );
 }
